src/validations/validators/email/email-validation.ts
import { InvalidFieldError } from '../../errors/invalid-field-error';
import { FieldValidation } from '../../protocols/field-validation';

export class EmailValidation implements FieldValidation {
  constructor(readonly field: string) {}

  validate(value: string): Error | null {
    const emailRegex = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/;

    return !value || emailRegex.test(value) ? null : new InvalidFieldError(this.field);
  }
}
